# bin/src/swagger.yaml
swagger: "2.0"
info:
  description: "Esta é uma api para controle de estados e suas cidades"
  version: "1.0.0"
  title: "CidadeEstadoAPI"
tags:
- name: "estados"
  description: "Controle dos estados"
- name: "cidades"
  description: "Controle das cidades"
- name: "users"
  description: "Controle dos usuários"
- name: "authentication"
  description: "Autenticao por token" 
paths:
  /estados:
    post:
      tags:
      - "estados"
      summary: "Adiciona um novo estado"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Estado que será adicionado"
        required: true
        schema:
          $ref: "#/definitions/estado"
      responses:
        400:
          description: "Entrada inválida"
        401:
          description: "Criado com sucesso"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    put:
      tags:
      - "estados"
      summary: "Atualiza um estado"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Estado a ser atualizado"
        required: true
        schema:
          $ref: "#/definitions/estado"
      responses:
        400:
          description: "Entrada inválida"
        201:
          description: "Criado com sucesso"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    get:
      tags:
      - "estados"
      summary: "Lista todos os estados"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
  /estados/{id}:
    get:
      tags:
      - "estados"
      summary: "Seleciona um estado"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        type: string
        name: id
        required: true
        
      responses:
        200:
          description: "Encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    delete:
      tags:
      - "estados"
      summary: "Exclui um estado"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        type: string
        name: id
        required: true
      responses:
        200:
          description: "Excluido"
        404:
          description: "Não encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
      
  /cidades:
    post:
      tags:
      - "cidades"
      summary: "Adiciona uma nova cidade"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Cidade que será adicionada"
        required: true
        schema:
          $ref: "#/definitions/cidade"
      responses:
        400:
          description: "Entrada inválida"
        401:
          description: "Criado com sucesso"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    put:
      tags:
      - "cidades"
      summary: "Atualiza uma cidade"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Cidade a ser atualizada"
        required: true
        schema:
          $ref: "#/definitions/cidade"
      responses:
        400:
          description: "Entrada inválida"
        201:
          description: "Criado com sucesso"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    get:
      tags:
      - "cidades"
      summary: "Lista todos as cidades"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
  /cidades/{id}:
    get:
      tags:
      - "cidades"
      summary: "Seleciona uma cidade"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        type: string
        name: id
        required: true
        
      responses:
        200:
          description: "Encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    delete:
      tags:
      - "cidades"
      summary: "Exclui uma cidade"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        type: string
        name: id
        required: true
      responses:
        200:
          description: "Excluido"
        404:
          description: "Não encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
  
  /users:
    post:
      tags:
      - "users"
      summary: "Adiciona um novo usuario"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Usuario que será adicionado"
        required: true
        schema:
          $ref: "#/definitions/user"
      responses:
        400:
          description: "Entrada inválida"
        401:
          description: "Criado com sucesso"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    put:
      tags:
      - "users"
      summary: "Atualiza um usuario"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Usuario a ser atualizado"
        required: true
        schema:
          $ref: "#/definitions/user"
      responses:
        400:
          description: "Entrada inválida"
        201:
          description: "Criado com sucesso"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    get:
      tags:
      - "users"
      summary: "Lista todos os usuarios"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
  /users/{id}:
    get:
      tags:
      - "users"
      summary: "Seleciona um usuario"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        type: string
        name: id
        required: true
        
      responses:
        200:
          description: "Encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
    delete:
      tags:
      - "users"
      summary: "Exclui um usuario"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        type: string
        name: id
        required: true
      responses:
        200:
          description: "Excluido"
        404:
          description: "Não encontrado"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
      
  /auth/register:
    post:
      tags:
      - "authentication"
      summary: "Gera token para usuario"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Usuario que receberá o token"
        required: true
        schema:
          $ref: "#/definitions/user"
      responses:
        400:
          description: "Entrada inválida"
        401:
          description: "Registrado com sucesso"
        500:
          description: "Erro no servidor"
  /auth/me:
    get:
      tags:
      - "authentication"
      summary: "Recupera informacoes do usuario autenticado"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        400:
          description: "Entrada inválida"
        401:
          description: "Encontrado com sucesso"
        500:
          description: "Erro no servidor"
      security:
      - api_key: []
  /auth/login:
    post:
      tags:
      - "authentication"
      summary: "Loga um usuario"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Usuario a ser atualizado"
        required: true
        schema:
          $ref: "#/definitions/user"
      responses:
        200:
          description: "Entrada inválida"
        201:
          description: "Logado com sucesso"
        500:
          description: "Erro no servidor"
  /auth/logout:
    get:
      tags:
      - "authentication"
      summary: "Desloga um usuario"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        201:
          description: "Deslogado com sucesso"
      security:
      - api_key: []
      
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
definitions:
  estado:
    type: "object"
    properties:
      id:
        type: "string"
      nome:
        type: "string"
      abreviacao:
        type: "string"
  cidade:
    type: "object"
    properties:
      id:
        type: "string"
      nome:
        type: "string"
      idEStado:
        type: "string"
  user:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"